refactor: split docs Rust modules
