+++ crates/ox_content_docs/src/extractor.rs
@@ -1,10 +1,9 @@
 //! Documentation extraction from source code using OXC parser.
 
-use ox_jsdoc::decoder::nodes::comment_ast::{LazyJsdocTag, LazyJsdocTagBody};
-use ox_jsdoc::parser::{
-    parse_batch_to_bytes as parse_jsdoc_batch_to_bytes, BatchItem as JsdocBatchItem,
-    ParseOptions as JsdocParseOptions,
-};
+mod jsdoc;
+mod jsdoc_tags;
+mod model;
+
 use oxc_allocator::Allocator;
 use oxc_ast::ast::{
     BindingPattern, Class, Comment, Declaration, ExportDefaultDeclarationKind, Expression,
@@ -15,233 +14,18 @@ use oxc_ast::ast::{
 use oxc_ast_visit::{walk, Visit};
 use oxc_parser::Parser;
 use oxc_span::{GetSpan, SourceType};
-use rustc_hash::{FxBuildHasher, FxHashMap, FxHashSet};
-use serde::{Deserialize, Serialize};
+use rustc_hash::FxHashMap;
 use std::path::Path;
-use thiserror::Error;
 
 #[allow(unused_imports)]
 use crate::profile_span;
 use crate::string_builder::{join2, join3, join5, StringBuilder};
 
-/// Result type for extraction operations.
-pub type ExtractResult<T> = Result<T, ExtractError>;
-
-/// Errors during documentation extraction.
-#[derive(Debug, Error)]
-pub enum ExtractError {
-    /// IO error.
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-
-    /// Parse error.
-    #[error("Parse error: {0}")]
-    Parse(String),
-
-    /// Unsupported file type.
-    #[error("Unsupported file type: {0}")]
-    UnsupportedFile(String),
-}
-
-/// Documentation item extracted from source code.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DocItem {
-    /// Item name.
-    pub name: String,
-    /// Item kind (function, class, interface, etc.).
-    pub kind: DocItemKind,
-    /// Documentation comment (JSDoc).
-    pub doc: Option<String>,
-    /// Source file path.
-    pub source_path: String,
-    /// Line number in source.
-    pub line: u32,
-    /// End line number in source.
-    pub end_line: u32,
-    /// Column number in source.
-    pub column: u32,
-    /// Raw JSDoc comment content without the outer delimiters.
-    pub jsdoc: Option<String>,
-    /// Whether the item is exported.
-    pub exported: bool,
-    /// Type signature (if applicable).
-    pub signature: Option<String>,
-    /// Extended base class/interface names.
-    #[serde(default)]
-    pub extends: Vec<String>,
-    /// Implemented interface names.
-    #[serde(default)]
-    pub implements: Vec<String>,
-    /// Whether a function/method declaration carries an implementation body.
-    /// `false` for overload signatures and ambient (`declare` / `.d.ts`)
-    /// declarations, and for non-callable items. Used to hide the implementation
-    /// signature when grouping overloads on TypeDoc symbol pages.
-    #[serde(default)]
-    pub has_body: bool,
-    /// Whether the item is optional.
-    #[serde(default)]
-    pub optional: bool,
-    /// Whether the item is readonly.
-    #[serde(default)]
-    pub readonly: bool,
-    /// Whether the item is static.
-    #[serde(default)]
-    pub r#static: bool,
-    /// Parameters (for functions/methods).
-    pub params: Vec<ParamDoc>,
-    /// Return type (for functions/methods).
-    pub return_type: Option<String>,
-    /// Members of an inline object literal return type.
-    #[serde(default)]
-    pub return_members: Vec<DocItem>,
-    /// Child items (for classes, modules, etc.).
-    pub children: Vec<DocItem>,
-    /// JSDoc tags.
-    pub tags: Vec<DocTag>,
-    /// Declaration type parameters (`<T extends C = D>`), in declaration order.
-    #[serde(default)]
-    pub type_parameters: Vec<TypeParamDoc>,
-}
-
-/// Parameter documentation.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ParamDoc {
-    /// Parameter name.
-    pub name: String,
-    /// Parameter type.
-    pub type_annotation: Option<String>,
-    /// Whether the parameter is optional.
-    pub optional: bool,
-    /// Default value (if any).
-    pub default_value: Option<String>,
-    /// Description from JSDoc @param tag.
-    pub description: Option<String>,
-}
-
-/// Type parameter documentation (`<T extends C = D>`).
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TypeParamDoc {
-    /// Type parameter name (e.g. `T`).
-    pub name: String,
-    /// Constraint after `extends`, when present.
-    pub constraint: Option<String>,
-    /// Default type after `=`, when present.
-    pub default: Option<String>,
-    /// Description merged from a `@typeParam` / `@template` tag (TSDoc).
-    pub description: String,
-}
-
-#[derive(Debug, Clone)]
-struct FunctionTypeMetadata {
-    params: Vec<ParamDoc>,
-    return_type: Option<String>,
-    return_members: Vec<DocItem>,
-    type_parameters: Vec<TypeParamDoc>,
-}
-
-impl FunctionTypeMetadata {
-    fn as_reference_metadata(&self) -> Self {
-        Self {
-            params: self
-                .params
-                .iter()
-                .map(|param| ParamDoc {
-                    name: param.name.clone(),
-                    type_annotation: param.type_annotation.clone(),
-                    optional: param.optional,
-                    default_value: param.default_value.clone(),
-                    description: None,
-                })
-                .collect(),
-            return_type: self.return_type.clone(),
-            return_members: self.return_members.clone(),
-            type_parameters: Vec::new(),
-        }
-    }
-}
-
-/// JSDoc tag.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DocTag {
-    /// Tag name (e.g., "param", "returns", "example").
-    pub tag: String,
-    /// Tag value.
-    pub value: String,
-    /// JSDoc type annotation, when the tag has a `{type}` part.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub type_annotation: Option<String>,
-    /// JSDoc name, when the tag has a name part.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    /// Whether the named part was marked optional.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-    /// Default value from `[name=value]` syntax.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub default_value: Option<String>,
-    /// Structured tag description parsed by `ox_jsdoc`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-}
-
-impl DocTag {
-    fn new(tag: String, value: String) -> Self {
-        Self {
-            tag,
-            value,
-            type_annotation: None,
-            name: None,
-            optional: None,
-            default_value: None,
-            description: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct ParsedParamTag {
-    name: String,
-    type_annotation: Option<String>,
-    optional: bool,
-    default_value: Option<String>,
-    description: Option<String>,
-}
-
-/// Kind of documentation item.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum DocItemKind {
-    /// Module or namespace.
-    Module,
-    /// Function.
-    Function,
-    /// Class.
-    Class,
-    /// Interface (TypeScript).
-    Interface,
-    /// Type alias.
-    Type,
-    /// Enum.
-    Enum,
-    /// Variable or constant.
-    Variable,
-    /// Class method.
-    Method,
-    /// Class property.
-    Property,
-    /// Constructor.
-    Constructor,
-    /// Getter.
-    Getter,
-    /// Setter.
-    Setter,
-    /// Enum member.
-    #[serde(rename = "enumMember")]
-    EnumMember,
-    /// TypeScript index signature.
-    #[serde(rename = "indexSignature")]
-    IndexSignature,
-}
+use self::jsdoc::{build_jsdoc_cache, parse_jsdoc_payload, ParsedJsdoc, MODULE_MARKER_TAGS};
+pub use self::model::{
+    DocItem, DocItemKind, DocTag, ExtractError, ExtractResult, ParamDoc, TypeParamDoc,
+};
+use self::model::{FunctionTypeMetadata, ParsedParamTag};
 
 /// Documentation extractor.
 pub struct DocExtractor {
@@ -439,138 +223,6 @@ impl Default for DocExtractor {
     fn default() -> Self {
         Self::new()
     }
-}
-
-/// Pre-parsed JSDoc data for one comment: `(raw, description, tags)`.
-type ParsedJsdoc = (String, String, Vec<DocTag>);
-
-/// JSDoc/TSDoc tags that mark a leading comment as a module/file comment.
-const MODULE_MARKER_TAGS: [&str; 3] = ["module", "packageDocumentation", "fileoverview"];
-
-/// Extract the JSDoc content body from a comment with a single allocation.
-fn extract_raw_jsdoc(comment: &Comment, source: &str) -> String {
-    let content = comment.content_span().source_text(source);
-    let trimmed = content.strip_prefix('*').unwrap_or(content);
-    trimmed.trim_matches('\n').to_string()
-}
-
-/// Parse a single JSDoc comment into `(raw, description, tags)` directly from its
-/// own span.
-///
-/// Unlike [`build_jsdoc_cache`], this does not key by `attached_to`, so it stays
-/// correct for comments that share an `attached_to` target — e.g. two leading
-/// file comments (`/** … @module */` followed by `/** @author … */`) that the
-/// parser both attaches to the first statement. Keying by `attached_to` would let
-/// the second comment overwrite the first in the cache.
-fn parse_jsdoc_payload(source: &str, comment: &Comment) -> ParsedJsdoc {
-    let raw = extract_raw_jsdoc(comment, source);
-    let items = [JsdocBatchItem {
-        source_text: comment.span.source_text(source),
-        base_offset: comment.span.start,
-    }];
-    let options = JsdocParseOptions { preserve_whitespace: true, ..JsdocParseOptions::default() };
-    let result = parse_jsdoc_batch_to_bytes(&items, options);
-
-    if result.diagnostics.is_empty() {
-        if let Ok(source_file) =
-            ox_jsdoc::decoder::source_file::LazySourceFile::new(&result.binary_bytes)
-        {
-            if let Some(Some(root)) = source_file.asts().next() {
-                let doc = root
-                    .description_text(false)
-                    .map_or_else(String::new, |description| description.trim().to_string());
-                // Module-entry parse happens once per file, so always format the
-                // tag value (cheap, and the raw path may surface it).
-                let tags =
-                    root.tags().map(|tag| DocVisitor::convert_jsdoc_tag(tag, true)).collect();
-                return (raw, doc, tags);
-            }
-        }
-    }
-
-    let (doc, tags) = DocVisitor::parse_jsdoc_fallback(&raw);
-    (raw, doc, tags)
-}
-
-/// Pre-parses every JSDoc comment in the program with a single batch call.
-///
-/// Oxc exposes comments separately from declarations, and the visitor resolves
-/// documentation by `attached_to` while walking the AST. Parsing each comment
-/// on demand repeats decoder setup and makes the visitor pay parse cost at
-/// every declaration. This cache batches decoder input once, stores the parsed
-/// `(raw, description, tags)` payload by `attached_to`, and lets the visitor do
-/// a cheap hash lookup in the hot declaration path. Comments that fail the
-/// batch parser still fall back individually so diagnostics do not poison the
-/// whole file.
-fn build_jsdoc_cache(
-    source: &str,
-    comments: &[Comment],
-    capture_raw: bool,
-) -> FxHashMap<u32, ParsedJsdoc> {
-    profile_span!("docs::parse_jsdoc");
-    let jsdoc_comments: Vec<&Comment> =
-        comments.iter().filter(|comment| comment.is_jsdoc()).collect();
-    if jsdoc_comments.is_empty() {
-        return FxHashMap::default();
-    }
-
-    let items: Vec<JsdocBatchItem<'_>> = jsdoc_comments
-        .iter()
-        .map(|comment| JsdocBatchItem {
-            source_text: comment.span.source_text(source),
-            base_offset: comment.span.start,
-        })
-        .collect();
-
-    let options = JsdocParseOptions { preserve_whitespace: true, ..JsdocParseOptions::default() };
-    let result = parse_jsdoc_batch_to_bytes(&items, options);
-
-    let failed: FxHashSet<u32> =
-        result.diagnostics.iter().map(|diagnostic| diagnostic.root_index).collect();
-
-    let mut cache: FxHashMap<u32, ParsedJsdoc> =
-        FxHashMap::with_capacity_and_hasher(jsdoc_comments.len(), FxBuildHasher);
-
-    if let Ok(source_file) =
-        ox_jsdoc::decoder::source_file::LazySourceFile::new(&result.binary_bytes)
-    {
-        for (index, (comment, root)) in jsdoc_comments.iter().zip(source_file.asts()).enumerate() {
-            let (raw, doc, tags) = match root {
-                Some(root) if !failed.contains(&(index as u32)) => {
-                    let doc = root
-                        .description_text(false)
-                        .map_or_else(String::new, |description| description.trim().to_string());
-                    let tags = root
-                        .tags()
-                        .map(|tag| DocVisitor::convert_jsdoc_tag(tag, capture_raw))
-                        .collect();
-                    // The decoder gave us the description/tags directly, so the
-                    // raw text is only needed when the caller will surface it.
-                    let raw = capture_raw.then(|| extract_raw_jsdoc(comment, source));
-                    (raw.unwrap_or_default(), doc, tags)
-                }
-                _ => {
-                    // The fallback parser works off the raw text, so we always
-                    // build it here; keep it only when the caller captures raw.
-                    let raw = extract_raw_jsdoc(comment, source);
-                    let (doc, tags) = DocVisitor::parse_jsdoc_fallback(&raw);
-                    (if capture_raw { raw } else { String::new() }, doc, tags)
-                }
-            };
-            cache.insert(comment.attached_to, (raw, doc, tags));
-        }
-    } else {
-        for comment in &jsdoc_comments {
-            let raw = extract_raw_jsdoc(comment, source);
-            let (doc, tags) = DocVisitor::parse_jsdoc_fallback(&raw);
-            cache.insert(
-                comment.attached_to,
-                (if capture_raw { raw } else { String::new() }, doc, tags),
-            );
-        }
-    }
-
-    cache
 }
 
 /// AST visitor for extracting documentation.
@@ -787,160 +439,6 @@ impl<'a> DocVisitor<'a> {
             .and_then(|stem| stem.to_str())
             .unwrap_or("module")
             .to_string()
-    }
-
-    fn convert_jsdoc_tag(tag: LazyJsdocTag<'_>, capture_value: bool) -> DocTag {
-        let tag_name = tag.tag().value().to_string();
-        // `value` is the formatted tag body that the raw `extract_file_docs`
-        // path surfaces. Normalize ignores it for param-family tags (it reads
-        // the structured name/type/description instead), so skip the costly
-        // `format_jsdoc_tag_value` join when the caller will not read it.
-        let value = if capture_value || !matches!(tag_name.as_str(), "param" | "arg" | "argument") {
-            Self::format_jsdoc_tag_value(&tag_name, &tag)
-        } else {
-            String::new()
-        };
-        let type_annotation = tag
-            .raw_type()
-            .map(|raw_type| raw_type.raw().trim().to_string())
-            .filter(|value| !value.is_empty());
-        let name =
-            tag.name().map(|name| name.raw().trim().to_string()).filter(|value| !value.is_empty());
-        let optional = name.as_ref().map(|_| tag.optional());
-        let default_value = tag.default_value().map(str::trim).filter(|value| !value.is_empty());
-        let description =
-            Self::format_structured_tag_description(&tag_name, &tag, type_annotation.as_deref());
-
-        DocTag {
-            tag: tag_name,
-            value,
-            type_annotation,
-            name,
-            optional,
-            default_value: default_value.map(str::to_string),
-            description,
-        }
-    }
-
-    fn format_structured_tag_description(
-        tag_name: &str,
-        tag: &LazyJsdocTag<'_>,
-        type_annotation: Option<&str>,
-    ) -> Option<String> {
-        if matches!(tag_name, "returns" | "return") {
-            let raw_body = tag.raw_body().map(str::trim).filter(|value| !value.is_empty())?;
-            let without_type = type_annotation
-                .and_then(|type_annotation| {
-                    let type_prefix = join3("{", type_annotation, "}");
-                    raw_body.strip_prefix(&type_prefix).map(str::trim_start)
-                })
-                .unwrap_or(raw_body);
-            return Self::clean_tag_description(without_type);
-        }
-
-        tag.description().and_then(Self::clean_tag_description)
-    }
-
-    fn format_jsdoc_tag_value(tag_name: &str, tag: &LazyJsdocTag<'_>) -> String {
-        if !matches!(tag_name, "param" | "arg" | "argument") {
-            if let Some(raw_body) = tag.raw_body().map(str::trim).filter(|value| !value.is_empty())
-            {
-                return raw_body.to_string();
-            }
-        }
-
-        let mut parts = Vec::new();
-
-        if let Some(raw_type) = tag.raw_type() {
-            let raw_type = raw_type.raw().trim();
-            if !raw_type.is_empty() {
-                parts.push(join3("{", raw_type, "}"));
-            }
-        }
-
-        if let Some(name) = tag.name() {
-            let name = name.raw().trim();
-            if !name.is_empty() {
-                let name = if tag.optional() {
-                    tag.default_value().map_or_else(
-                        || join3("[", name, "]"),
-                        |default_value| join5("[", name, "=", default_value, "]"),
-                    )
-                } else {
-                    name.to_string()
-                };
-                parts.push(name);
-            }
-        }
-
-        if let Some(description) =
-            tag.description().map(str::trim).filter(|value| !value.is_empty())
-        {
-            if parts.is_empty() {
-                parts.push(description.to_string());
-            } else {
-                parts.push(join2("- ", description));
-            }
-        }
-
-        if !parts.is_empty() {
-            return parts.join(" ");
-        }
-
-        if let Some(raw_body) = tag.raw_body().map(str::trim).filter(|value| !value.is_empty()) {
-            return raw_body.to_string();
-        }
-
-        tag.body().map_or_else(String::new, |body| match body {
-            LazyJsdocTagBody::Generic(body) => body.description().unwrap_or_default().to_string(),
-            LazyJsdocTagBody::Raw(body) => body.raw().to_string(),
-            LazyJsdocTagBody::Borrows(_) => String::new(),
-        })
-    }
-
-    /// Fallback parser used when the external JSDoc parser cannot produce a root.
-    fn parse_jsdoc_fallback(comment: &str) -> (String, Vec<DocTag>) {
-        let mut description_lines = Vec::new();
-        let mut tags = Vec::new();
-        let mut current_tag: Option<(String, Vec<String>)> = None;
-
-        let lines: Vec<String> = comment
-            .lines()
-            .map(|line| {
-                let trimmed = line.trim_start();
-                let trimmed = trimmed.strip_prefix('*').unwrap_or(trimmed);
-                trimmed.strip_prefix(' ').unwrap_or(trimmed).trim_end().to_string()
-            })
-            .collect();
-
-        for line in lines {
-            let trimmed = line.trim_start();
-            if let Some(without_at) = trimmed.strip_prefix('@') {
-                // Save previous tag if any
-                if let Some((tag, value_lines)) = current_tag.take() {
-                    tags.push(DocTag::new(tag, value_lines.join("\n").trim().to_string()));
-                }
-
-                let split_at = without_at
-                    .char_indices()
-                    .find_map(|(index, ch)| ch.is_whitespace().then_some(index))
-                    .unwrap_or(without_at.len());
-                let tag_name = without_at[..split_at].to_string();
-                let tag_value = without_at[split_at..].trim_start().to_string();
-                current_tag = Some((tag_name, Vec::from([tag_value])));
-            } else if let Some((_, ref mut value_lines)) = current_tag {
-                value_lines.push(line);
-            } else {
-                description_lines.push(line);
-            }
-        }
-
-        // Save last tag if any
-        if let Some((tag, value_lines)) = current_tag {
-            tags.push(DocTag::new(tag, value_lines.join("\n").trim().to_string()));
-        }
-
-        (description_lines.join("\n").trim().to_string(), tags)
     }
 
     fn format_type_parameter_declaration<T>(
